requirements.txt
Django==3.1.5
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pytest-django==4.1.0
pytest==6.2.1
gunicorn==20.0.4
dj-database-url==0.5.0
whitenoise==5.2.0
pytest-cov==2.10.1
flake8===3.8.4
black==20.8b1
isort==5.7.0